linux-grub-setup: force update-grub when used
If you happen to have removed an old kernel while also
installing a new one you will end up selecting a wrong
kernel with grub-set-default. So ensure you always
update your grub.cfg before running grub-set-default.

This would typically be done by the distro install.sh
script at install time, but well, if we really mess up
the selection we'll end up with a wrong kernel, or worst
a recovery kernel. So it is not a big issue to run this
twice.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>