fstests: do not run large disk tests by default
generic/048 was reported as hung test, but the root cause for that
was that qemu got OOM killed when running the test.

generic/048 happens to be listed as a "large disk" test because
it requres 10G of scratch space.

There is already another test from the "large disk" test that is listed
as hung test (generic/320).

These tests allegedly generate memory pressure in highly concurrent
test eviroment, so to reduce noise, disable those tests by default.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>