kernel-ci: implement incremental steady-state goal
With this configuration, make fstests baseline-loop will resume
running the loop from the last success counter instead of resetting
the counter.

This methodology makes sense if all the failed testes encountered
on failure have been added to the expunge list. In that case, the
remaining tests in the baseline did actually complete the incremental
number of run loops without any failure.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>