fstests: use expunge lists from baseline version
When guests are running a kernel with localversion suffix to
uname, use the base version to get the expunge lists.

For example, when testing kernel version 5.10.109-xfs-1,
which is based on kernel 5.10.109 plus xfs backports, the
expunge list of 5.10.109 baseline should be used as we are
not going to create a baseline from this version.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>