reboot-limit: add a systemd-analyze clutch at bootup
If you run systemd-analyze at bootup it may fail with:

$ sudo systemd-analyze
Bootup is not yet finished (org.freedesktop.systemd1.Manager.FinishTimestampMonotonic=0).
Please try again later.
Hint: Use 'systemctl list-jobs' to see active jobs

When running the reboot-limit test demo workflow to test to see
how many boots can happen without failure on a system we need
then to add a wait-for-boot-to-complete clutch before we run
the systemd-analyze (when that is enabled).

We can accomplish this with:

systemctl is-system-running --wait

So add that if before we collect systemd-analyze results.
This is only applicable when CONFIG_REBOOT_LIMIT_ENABLE_SYSTEMD_ANALYZE=y.
Later on we *may* want to just use this and remove the silly
arbitrary delay on bootup post_reboot_delay. Using this
$(systemctl is-system-running --wait) may be nice for complex nodes
where bootup may take a while, otherwise we're just rebooting *as*
soon as the post_reboot_delay completes and perhaps that may do
some insane things.

But since CONFIG_REBOOT_LIMIT_ENABLE_SYSTEMD_ANALYZE=y is the default
today we simpley force using $(systemctl is-system-running --wait)
systemd-analyze is going to be used, that is always by default today
anyways.

We can later make it a strong requirement all the time even if
you don't have CONFIG_REBOOT_LIMIT_ENABLE_SYSTEMD_ANALYZE=y, but
for now this makes sense as-is as we do want to try to speed up
reboots as much as possible.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>