fstests: run the 'auto' tests group by default
This is the recommended set of tests for regression tests
or to be more precise, tests that are NOT in the auto group
are unlikely good regression tests.

oscheck helpers for older kernel distros use this skip list for xfs:
"tape clone dedupe dax dangerous_repair dangerous_online_repair broken"

Most of the groups in this skip list are irrelevant when running the
'auto' tests group, because there is no 'dangerous*' nor 'broken' test
in the 'auto' group.

There is only one 'auto tape' test and if this is not an adequate
regression test it should be removed from the 'auto' group.

The tests in the 'clone dedupe dax' groups are going to be skipped
automatically besed on dax/reflink support, so there is no need to
explicitly skip test and loose potential test coverage.

The skip list for xfs on debian/testing has already been removed by an
earlier commit.  The skip list for xfs on distros with older kernel
versions were left unchanged, because I did not test those distros and
there may be a kernel version (i.e. < v4.16), where reflink are
supported but still experimental.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>