scripts/gen_hosts.sh: skip extra nodes when not needed
When CONFIG_KDEVOPS_BASELINE_AND_DEV is not enabled the ansible
hosts file should not have the hostfiles with the -dev postfix.
This is not an issue when a target workflow uses blktests or fstests
as the scripts scripts/workflows/*/*/gen_manual_hosts.sh already
check for CONFIG_KDEVOPS_BASELINE_AND_DEV.

This only fixes the ansible hosts file when a workflow like
reboot-limit is used, or no workflow is selected, and you don't
want to bother with a -dev extra node.

This doesn't fix any known reported buts. I ran into this
through manual inspection.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>