fstests: do not report the None test as Hung-Stalled
My loops keeps getting terminated shortly after restart with a status of
a test named None which is Hung-Stalled.

When I ssh the the so called hung guest I see that oscheck is runnings
and everything looks sane.

Report this case as Timeout instead of Hung-Stalled.

Also increase the ssh communication timeout to 120s which proved to
avoid this timeout condition.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>