xfs: add generic/600 as failing in v5.10.y
It failed after more than 100 runs.

This looks like a test bug related to 1 second drift on clock change:
"set grace to 1653569363 but got grace 1653569362"
that upstream attempted to fix in:
4f9dd735 generic/60[01]: fix test failure when setting new grace limit

The test is alredy expunged in ext4 5.17 and other xfs configs,
so that test should probably go to maybe-broken list.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>