btrfs: failures for v5.17-rc7 when moving from btrfs to xfs on the host
These are new failures for v5.17-rc7 after we moved from the host
using btrfs to xfs on the guest nvme drives. The scripts below
are used to help simplify things and complete the baseline faster:

./scripts/workflows/fstests/find-common-failures.sh -l workflows/fstests/expunges/5.17.0-rc7/btrfs/unassigned/
./scripts/workflows/fstests/remove-common-failures.sh workflows/fstests/expunges/5.17.0-rc7/btrfs/unassigned/

Is this sort of change possible to reveal new issues? Well it turns
out it is! Because we ran into the same exact thing when testing
XFS on v5.10.105. Refer to kdevops commmit 354789bd6572 ("xfs: fix
comment about xfsprogs regression on v5.10.105").

So in this case the old test environment was using:

CONFIG_WORKFLOW_DATA_FSTYPE="btrfs"
CONFIG_FSTESTS_SPARSE_FSTYPE="btrfs"

Now the host is are using:

CONFIG_WORKFLOW_DATA_FSTYPE="xfs"
CONFIG_FSTESTS_SPARSE_FSTYPE="xfs"

Which means we should *also* test btrfs on ext4 as well.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>