fstests: copy over *.full files on failures
When a failure is detected we copy now only the *.bad files
and *.dmesg file but we don't copy over the respective *.full file
copy this file over.

This now completes collecting all artifacts for a test falures.
If we wanted to take a second step to add these files to a report
we can do that now, all the pieces are there now for to *start*
a bug report.

An acual bug report requires analysis of the results and careful
inspection of what the data tells us.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>