fstests/augment_expunge_list.py: add support for processing dmesg files
The script which augments our expunge lists checks for *.bad files
only. Add support for processing *.dmesg files as well. The bad
files have the form:

  * btrfs/123.out.bad

while dmesg files have the form:

  * btrfs/123.dmesg

and so we just need to conside the first part of the file
after we split the filename string with ".".

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>