kdevops: add initial qemu build support and add a CXL enhanced repo
This adds initial build support for qemu so to enable working
with technology which is not available yet on the versions of qemu
carried by distributions yet.

Examples for this is CXL support which does not yet have
support through a qemu package on most distributions, so
the only thing you can do is build qemu from scratch.

This adds the basic initial distribution support for all the
distributions we currently support. This just let's us build
qemu and install it.

The next steps will be to add kconfig logic to take advatage
of some bleeding edge tech which is still under development
such as CXL.

Building qemu from source is obviously useful beyond CXL, this
is just an example tech which is not yet upstream.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>