btrfs: add generic/371 as having large runtime variability
generic/371 can sometimes takes up to 10 times it typically takes
to run on btrfs. This variabilty is odd and should be looked into. The
kdevops fstest watchdog is configured to pick this up, it detected this
variability as follows:

Hostname              Test-name   Completion %  runtime(s) last-runtime(s)   Stall-status Kernel
linux517-btrfs-simple generic/371 1173%         305        26                Hung-Stalled 5.17.0-rc7

The stall detection is part of the kdevops fstests watchdog, it detected
it took more than 10 times to run the test once than the first time it
ran it once the kernel-ci loop kicked off. We can bump the default max
multiplier (CONFIG_FSTESTS_WATCHDOG_HUNG_MULTIPLIER_LONG_TESTS=10) but
first we should think if this makes sense.

We should change the "Hung-Stalled" to "Runtime-variability" or something
like that to make this clearer for reports as well.

It may very well be that for this test it makes sense for a test to
take 10 times it typically takes, in such cases then we need to add
semantics for that.

The way generic/371 works is it uses two separate writes in parallel
to two separate files, one using write, and the other fallocate:

echo "Silence is golden"
for ((i=0; i<100; i++)); do
        $XFS_IO_PROG -fc "pwrite 0 80M" $testfile1 >/dev/null
        rm -f $testfile1
done &
pids=$!

for ((i=0; i<100; i++)); do
        $XFS_IO_PROG -fc "falloc 0 80M" $testfile2 >/dev/null
        rm -f $testfile2
done &
pids="$pids $!"
wait $pids

I can't think why this would diverge so much in run time, is this
acceptable? For now I think it is fair to say no, and expunge this
for all test sections.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>