build_qemu: fix install target
The qemu build process will complain if you follow the typical:

./configure
make
sudo make install

The reason is it relies on git submodules and by default if the
user is different at install time, it will try to refresh the
git submodules and there is a persmission check so that if the
user at this step is different than the origianl one it will
complain.

As per the README the way the way to address this is to use:

./configure --with-git-submodules=validate

But after this we must also run:

./scripts/git-submodule.sh update

Fix this.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>