build_qemu: add an optional directory for localhost builds
qemu is the first component which we optionaly allow building
locally on kdevops onto the host. The assumption before was
that we were only building things on the guest, and used
{{ data_path }} for that. But since that is for the target
nodes / guests it is incorrect to use that. So we need to
define a new diretory for local builds. This directory
is not needed to be exposed as an option if you are not
building qemu, and so let's not confuse users with this
and only expose it *iff* the user is building qemu.

We add the CONFIG_NEEDS_LOCAL_DEVELOPMENT_PATH bool to let
other future kconfig entries which may need this variable
as well. They would just need to select it as we do now with
CONFIG_QEMU_BUILD.

Fixes: ecabd29de4c9 ("kdevops: add initial qemu build support and add a CXL enhanced repo")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>