fstests: migrate .kdevops_fstests_setup install to ansible
The install of the /root/.kdevops_fstests_setup file was originally
installed using the old Makefile from oscheck. We can simplfiy
this greatly by using ansible.

Do that and provide a new target to enable to test this specific series
of kdevops tasks as the target:

  make fstests-kdevops-setup

Just running 'make fstests' or any of the test specific targets will
always run this as well. We do this so that if you re-configure
your test you can easily propagate the changes.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>