fstests: replace gen_fstests_config.sh with ansible
The gen_fstests_config.sh script is currently used to generate
a target node filesystem configuration file for the target filesystem.
This is a bit fragile and complex for a few reasons:

 * If we update .config we need to ensure we manually re-generate it
   and copy it over using 'make fstests' which also does a full rebuild
   of fstests.
 * The sed stuff is long / complex and error prone

We can replace this by using ansible and extend the variables on
the extra_vars.yaml file and simply use the ansible template
module to transcribe the result with our ansible variables.

We also then *always* regenerate it when running each test which
ensures we are paranoid and have a fidelity with the exact
configuration at each run time.

We might be able to use ansible later to also convert the
other generation files which used sed such as:

  * scripts/gen_nodes_file.sh
  * scripts/workflows/fstests/*/gen_manual_hosts.sh
  * scripts/workflows/blktests/gen_manual_hosts.sh
  * scripts/gen_hosts.sh
  * scripts/gen_nodes_file.sh
  * scripts/gen_ssh_key.sh
  * scripts/gen_tfvars.sh

We'd just use a localhost ansible role.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>