fstests: ensure extra_vars.yaml is updated on updates to roles
If you provide an update to an ansible role and if you want to make use
of a new variales you would have to update the respective ansible role
defaults/main.yml file. We would have a Kconfig option to enable
variability. However, today we only pass on the filesystem configuration
file when we run 'make install' once. Fixing so that we send this file
updated using the current methodology doesn't scale well with all the
sed stuff.

Instead, start first by ensuring the extra_vars.yaml file will be
updated if we update an ansible role with a new variable. The next
patch will address ensuring we can update the filesystem configuration
in a scalable way *always*.

Signed-off-by: XFS shared account <xfs@bgt140507bm04>