fstests: add fs configuration for root's .kdevops_fstests_setup
We already have support to run fstests through ansible but often
times we want to log in and run fstest's check.sh manually. Allow
for this by parsing the filesystem configuration as check.sh does.

We do slight modifications to the routines used to reflect a small
delta is incurred to support this. We may just want to upstream adding
suport for adding this file to fstests upstream to root's home directory
as we do, but that will require some discussion.

With this in place if you want to run ./check.sh manually all you'd
have to do is just log in, cd /var/lib/xfstest and run first, for
instance:

cd /var/lib/xfstests/
./gendisks.sh -d -m
./check -s <foo> generic/002

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>