vagrant: make CPU memory configurable and make 2 GiB RAM default
The current kdevops vagrant guest default is to use 8 GiB or RAM
and 2 vCPUs. But prior testing shows we can pull of just 2 GiB of RAM
with 8 vCPUs just as well. The only known issue was with xfs_scrub
requiring more than 2 GiB of RAM but that was an issue that was supposed
to be fixed upstream. So just make 2 GiB the default guest RAM
specification and strive for this.

Compiling fstests also does not require much CPU resources.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>