sysctls: enable node sysctl overwriting of vm.overcommit_memory
It can be useful to overwrite the default sysctls on target
nodes when testing the kernel. This let's us overwrite these
in a generic fashion.

This starts by allowing us to overwite the value of vm.overcommit_memory.
This also documents a bit of tribal knowledge on uses cases on examples
where this might come in handy.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>