kdevops: make KDEVOPS_BASELINE_AND_DEV=n default now
Since its inception kdevops has been enabling instantiating two
nodes per target test. One node is used to target the baselin and
the secondary node can be used to help reproduce issues when found
or for development purposes. The assumption that users have enough
resources to spawn two nodes per test is however a bit excessive
as we start covering more filesystems and subsystems. For example
to test XFS fully on v5.10.y we require at least 10 nodes
(excluding the realtime configuration):

  * linux510-xfs-reflink-1024-dev
  * linux510-xfs-reflink-normapbt
  * linux510-xfs-nocrc-dev
  * linux510-xfs-nocrc-512-dev
  * linux510-xfs-nocrc-512
  * linux510-xfs-reflink-1024
  * linux510-xfs-nocrc
  * linux510-xfs-reflink-normapbt-dev
  * linux510-xfs-reflink
  * linux510-xfs-reflink-dev

Since we recently dropped the default RAM per node to 2 GiB this means
10 GiB for this kernel release. Enabling KDEVOPS_BASELINE_AND_DEV=y
would have meant extra ${NODE}-dev nodes so 20 GiB. Slightly before
this we defaulted to 8 GiB per node, so 160 GiB for all. This is why
we strive to go with the minium resources now.

If you try to cover XFS fully on about 3 kernel releases, say v5.10.y,
linus' latest and linux-next, that's about 30 GiB. But there are 10
kernel releases tracked on kernel.org as of today:

  1)  linux-next
  2)  5.19-rc1
  3)  5.18.y
  4)  5.17.y
  5)  5.15.y
  6)  5.10.y
  7)  5.4.y
  8)  4.19.y
  9)  4.14.y
  10) 4.9.y

So minimum, rougly 100 GiB for one filesystem for 10 kernel releases.
We have 3 main filesystems today in Linux:

  * xfs
  * btrfs
  * ext4

With the current defaults with KDEVOPS_BASELINE_AND_DEV=n we can strive
to tests most filesystems with about ~300 - 500 GiB (leaving wiggle room
for the configurations missing for the other filesystems).

This is back of napkin calculation for fstests. Similar math can be done
for blktests.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>