scripts/bringup_vagrant.sh: fix version check
The version check added breaks with newer vagrant-libvirt uses.
So make the check a bit smarter and check using math for the versions.
This also makes the check easier to read and let's us scale this if
needed in the future.

This and the prior vagrant commit makes this all work on the latest
and greatest debian-testing as of today. We really need to get a
continous integration testing going.

Fixes: e0a82ea04bc7 ("kdevops: use --no-parallel for broken version of vagrant-libvirt")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>