vagrant/Vagrantfile: add support for nested virtualization using vagrant
If you want to get nested virtualizationg going on Linux, say
using kdevops to first create a large beefy guest, and then inside that
guest another smaller guest you need take few steps:

1) On the host ensure nested virtualization is enabled on the respective intel
   or amd module using the nested module parameter. You can ensure this on
   bootup on the file /etc/modprobe.d/kvm.conf by having
   options kvm_amd nested=1
   or
   options kvm_intel nested=1

2) If you are instantiating a libvirt guest using vagrant and then
   instantiate another guest on that guest you will need to modify
   a few things on the guest.

The changes in this patch do the work for 2). It detects if we're
nested using systemd-detect-virtd. For now we just do something
different if not on bare-metal. It may not make sense to support
and modify this for each virt-type... not sure yet.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>