fstests: fix processing of *.full files with loop_control
Commit af873ad905f80 ("fstests: copy over *.full files on failures")
uses the list if bad files found to then do a search using the ansible
stat module, and *then* registers this. When looping over items ansible
let's you shorten the loop message with the loop_control option.

It would seem to be a bug in Ansible that if you put use an
item.stat.path on the loop_control but still ensured that the
file existed before using 'when: item.stat.exists' that ansible
will still complain should this file not be present.

To fix this we can either remove the loop_control call and let
ansiblew print a long line as it would wish if it found the item,
but we can also work around this limitation by instead relying
on the inner loop which was used to register the items, and so
item.item would give us the original item we used, and we can
simply define local variables using that.

While at it, fix the file being looked for by using split('/')
and split('.') for the directory and filename replacement we are
doing instead of using splitext as it is not doing what we want.

Now, if generic/230.out.bad exists we will look for the expected
generic/230.full instead of generic/230.out.full which is what
we were looking for before.

Likewise now if generic/230.out.bad exists and the expected
generic/230.full file does not exist things still work without
a failure. We just copy the file if it exists.

Reported-by: Amir Goldstein <amir73il@gmail.com>
Fixes: af873ad905f80 ("fstests: copy over *.full files on failures")
Fixes: 5ed18e348d31a ("fstests: fix copy of test full bad result file")
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>