kdevops: add ansible role to install make menuconfig dependencies
Often times you want to just git clone kdevops and immediately after
that on a new host you manually install the dependencies for running
'make menuconfig'. Add an ansible role which does just this so that
the only thing you really need to install to get running with kdevops
is:

  * ansible
  * make

Then you just run the following on any new system so you can get
the make menuconfig dependencies.

make menuconfig-deps

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>