kdevops: replace ansible hosts generation jinja2 templating
The way we were generating the top level ansible hosts file
doesn't scale easily for a few reasons:

  * it relies on sed, and eventually we'll run out of
    space in arguments for what we want to parse
  * sed variables cannot have underscores and makes it
    hard to read
  * the semantics for dedicated tests require trying to
    define what is a possible host in multiple places and
    this can easily get out of hand
  * we're not taking advantage of kconfig semantic correctness
    which ensures that hosts can only be allowed as per its
    requirements

To fix this use ansible's jinja2 template support. Ansible
uses a template module to let you parse files using the
template module using jinja2. For those of you not familiar
with jinja2 it is a template engine designed originally for
HTML templating. It is a good fit as we can replace the
horrendous eye-sores created by trying to work with sed
replacements and is supported by both Ansible and Salt.

The git diff stats says it all:

 18 files changed, 304 insertions(+), 399 deletions(-)

And this is just with 2 major comple test frameworks added.

Using jinja2 will let us scale tremendously. And generation
of the nodes yaml file is the next obvious target. That should
make kdevops extremely adaptable and scalable. A hint to the
future: we *also* now even have path forward for *real* dynamic
kconfig magic ;) -- if you wanted a way to dynamically create
Kconfig files for *any* system you are on for PCI passthrough
for example, *I think* this should get us there.

Getting to the point this is usable was pretty hard to get to,
this took time, patience and a bit of opera music but its all
there now.

[0] https://en.wikipedia.org/wiki/Jinja_(template_engine)

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>