fstests: add semantic gap for when FSTESTS_*_MANUAL_COVERAGE is disabled
When a Linux distribution doesn't *know* / *prefer* what sections it
should test for a filesystem then the respective filesystem options
are disabled. For instance we have the following kconfig options
which can be used to allow the user / distro defaults to enable or
disable certain filesystem test sections:

  * CONFIG_FSTESTS_XFS_MANUAL_COVERAGE
  * CONFIG_FSTESTS_BTRFS_MANUAL_COVERAGE
  * CONFIG_FSTESTS_EXT4_MANUAL_COVERAGE

Ideally, we want to be able to generate the ansible hosts file (./hosts)
for dedicated tests using test section coverage based on what you
enabled. This works well when enabling the manual coverage options
for a filesystem a then we have a respective filesystem kconfig option
per test section. For instance when CONFIG_FSTESTS_XFS_MANUAL_COVERAGE
is enabled we'll have in our resulting configuration .config file when
these are enabled:

  * CONFIG_FSTESTS_XFS_SECTION_CRC=y
  * CONFIG_FSTESTS_XFS_SECTION_NOCRC_512=y
  * CONFIG_FSTESTS_XFS_SECTION_REFLINK=y
  * CONFIG_FSTESTS_XFS_SECTION_REFLINK_1024=y
  * ... etc ...

If you disable CONFIG_FSTESTS_XFS_MANUAL_COVERAGE then we have to
keep a list of known sections somewhere and then use these to expand
on the hosts file. This is what we do today with the secript
scripts/workflows/fstests/xfs/gen_manual_hosts.sh.

But keeping a list of valid sections in two places does not scale.
We want to allow *growth* by just expanding on the Kconfig files
for each filesystem for a dedicated fstests setup. To enable this
we must close the gap and add kconfig options (also known as symbols)
for each filesystem test section for when the respective filesystem
tests manual symbols are disabled (CONFIG_FSTESTS_*_MANUAL_COVERAGE).

This fixes that semantic gap.

From a semantics / growth perpsective this was also the correct thing
to do becuase *if* we have filesystem section which had architecture
specific requirements (currently XFS does for FSTESTS_XFS_SECTION_BIGBLOCK
for powerpc / aarch64) we'd then have to add these dependency
annotations in two places, both in Kconfig and on the script which
generates the ansible hosts files.

Fixing this will allow us to unify use one simple algorithm for
determing what hosts to output later in subsequent patches:

  * What filesystem sections are available for this filesystem?
  * Of the above, now, which ones are enabled?

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>