fstests: add semantic gap for when BLKTESTS_MANUAL_COVERAGE is disabled
As we did with fstests we want to simplify and unify how we determine
what is a new test section by simply using Kconfig because:

  * this scales easily, as new tests sections are added
    upstream we just have to go and edit the Kconfig file
  * we then also allow kconfig to deal with the modeling
    variability constraints, for instance such as architecture
    constraints or requirements for the drives, an example is
    ZNS requires special drives.

This will then also allow us to us more easily generate the
ansible hosts file for dedicated blktests in subsequent patches.

This commit fills in the gap to ensure that the sections for
blktests are always specified on the .config file. The situation
is easier for blktests as we have one word per test section.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>