kdevops: generate nodes & Vagrantfile file with jinja2 templating
The file vagrant/kdevops_nodes.yaml is today generated using a set
of scripts which depend on what workflow / filesystem you want to
use. This is ugly and complex.

Likewise the top level vagrant/Vagrantfile is getting quite out of
hand with tons of bells and whistles added for each new technology
we want to add support for.

We can simplify things consirably by just using jinja2 templating,
so that the nodes file is generated based on input in a more
gramatically consistent manner similar to how we generate the
ansible hosts file. Later hope is to just share some of this
code between both.

The vagrant/Vagrantfile is complex because it supports both
virtualbox and libvirt, and it also supports a slew of technologies.
This will only grow more in the future. For example, we'll be
extending it soon with a slew of CXL options. To make things
easier to maintain and to allow this to be easier to grow and
read we now generate the vagrant/Vagrantfile by trimming it
with code which is not needed. This should make it easier for
users to read and understand the file while also extending it
with new features.

Later on we can make things even easier to read by just justing
include files. But take note: jinja2 processing has restrictions
on variables used, so you can't exactly be setting variables to
do complex math and expect them to be carried out on macros.
Hacks like using loop.index, etc are required [0]. If you
need anything complex, that should all be done as on the ansible
tasks variables, as ansible does allow you to do pretty complex
things with variables. Templating should be used for simple
processing. That's it.

We can still do better! For instance, now that we have the power
of stuffing variables within ansible, I don't think we need the
vagrant/kdevops_nodes.yaml file at all. We just have to add data
for simple loops to be processed. Then, in consideration of growth
for new technology, the approach to extend vagrant/kdevops_nodes.yaml
for new technology seems wrong, we should just have to process some
variables on ansible and then have the Vagrantfile have a macro to
process the new technolgy. For instanced, the fact that the code to
support ZNS in the vagrant/Vagrantfile is left in place whether or
not you support ZNS is just cumbersome. We should just define a new
variable for nvme drives and have a bool for zns drives. Then we
split the current nvme set into two arrays and have the template
generation work on two separate arrays using the shared macro. We
may then just want to create a separate PCI root bus for each, one
for regular nvme drives, and another for ZNS drives.

Adding CXL support should learn from these growing pains and lead
the way with cleaner processing.

And again the diff stat says it all:

 27 files changed, 476 insertions(+), 822 deletions(-)

Converting over to generate the required files for terraform using
jinja2 is next. And then we can have a path forward to grow in a very
nice way for both cloud and local virtualization.

[0] https://jinja.palletsprojects.com/en/3.1.x/templates/

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>