docs: update git subtree documentation and references
The ansible role update_ssh_config_vagrant shares a python script
with a terraform module we have pushed onto the terraform registry.
To enable sharing code we use a git subtree for the python code.
So the ansible role update_ssh_config_vagrant is one user and the
other is an external terraform module.

Updating this code can not be so trivial to grasp if you are looking
at it at first glance so make this clearer. Also update the top level
Makefile.subtrees so to explain how to push locally queued changes back
upstream.

Next I'll work on getting this code on the linux-kdevops organization
and update the docs for update_ssh_config code a bit so make it easy
to explain how to contribute code.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>