xfs: add xfs/291 as failing for all configs in 5.10.y
This is an xfs_repair test and it failed on another section with
"xfs_check failed" error on debina/buster with old xfsprogs.

It is one of the few tests that still runs xfs_check despite
the fact that FORCE_XFS_CHECK_PROG is not defined.
xfs_check is not considered reliable and xfs maintainer has
patched it out a long time ago, so its not worth the trouble.

Signed-off-by: Amir Goldstein <amir73il@gmail.com>