terraform: provide proper generation for nodes with jinja2
The way terraform currently figure out what nodes to bring
up is by parsing the vagrant/kdevops_nodes.yaml file using
a yaml. This requires us to use a yaml parser and download
that.. it's is a holy mess. The vagrant/kdevops_nodews.yaml
file should be kept just for vagrant uses, and in fact I
am not even sure we need it anymore.

So just generate our own nodes file for terraform using
the same ansible role, and add variable we can use for it,
we use all_generic_nodes.

Again, the diff stat says it all:

 29 files changed, 66 insertions(+), 263 deletions(-)

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>