terraform: bump binary requirement from 0.12.19 to 1.2.3
0.12.19 is ancient and is signed with an expired key.
Require the latest and greatest terraform which as of today
is 1.2.3.

The new terraform uses a lock file to ensure you don't
get confused and use a provider you did not intend to.
So when you run 'make mrproper' we ensure to kill it too.
The lock file is .terraform.lock.hcl.

While at it, we bump the providers to their latest code
as we had not done this in years. I at least tested
aws and that works. The rest I tested with::::

  * terraform init
  * terraform plan

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>