terraform: replace tfvars.tf generation with jinja2
OK the last the part of the jinja2 conversion was the
terraform tfvars file which is used to specify variables
for it's own template stuff.

That's it! We can now scale. And at least on the terrform
front we can probably optimize this further as there is
too many variables we have to pass back to terraform. It
would be idea if we can just have tarraform read
extra_vars.yaml file for its own template engine.

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>