vagrant: add support for the linux-kdevops images
We may want to test linux-next kernels with new
technology. Enable support for this. The first
build uses linux-next tag next-20220629.

To learn to build your own box:

https://github.com/linux-kdevops/kdevops/blob/master/docs/custom-vagrant-boxes.md

If you want to help with this effort with kernels from linus tree,
stable kernels, maintainer kernels, etc, you can be added to the
linux-kdevops org for vagrant to share same namespace.

The first image is this, it is debian based but the namespace
used can support other distros:

https://app.vagrantup.com/linux-kdevops/boxes/debian-next-20220629

Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>