kdevops: fix parallel builds when using make_num_jobs
Adam repoted long ago some odd build issues when building qemu.
The issues was that he had used 'make -jX' and this is not needed
on kdevops, we deal with parallelism for you. The subsequent failure
he found was a real build failure on upstream qemu, we can't do
anything about that.

Although at the time I did verify a few parallel make commands were
being issued, I think the system I used was also too small, it only
had about 8 cores so naturally it is easy to end up with 4 make
commands even if you are not running in parallel. It gave me the
incorrect impression that parallel builds are correct when using
make_num_jobs.

When testing building on a system with 48 cores, I can clearly
see the difference and the issue of parallel builds not happening
is evident to me now.

Although the upstream documentation provides an example using
NUM_THREADS: 4, this is completley bogus! All it is doing is
allowing you to pass on to the make command some arbitrary
arguments. So we're actually just running: gmake NUM_THREADS=4
which really does nothing!

As of ansible 2.0, there is a jobs argument we can use, but
the target node may have a different number of nodes than the
controller, and and node may have a different set of cores,
so there is no easy way to do this than by just open coding
running nproc first and then using the output. So fix this for
all users of the stupid make_num_jobs I had which was incorrectly
inspired by the bogus help page on ansible [0]. When someone has
time someone should just add an example to demo how we do it now
as this is just rediculous.

By default kdevops builds only a few things, the most intense thing
is the kernel but that was built using another method, and that worked.
Only the stuff using make_num_jobs which used the bogus NUM_THREADS
stuff were not being built in parallel. It is important to itermize
these though:

  * qemu (useful for folks doing CXL hacking or R&D storage development)
  * on suse - building nbd when using blktests
  * blktests:
    - blktrace
    - blktests
    - dbench
  * fstests:
   - dbench
   - fstests

In the future, if you want to really instpect number of parallel
builds processes use something like pstree and a system with high
number of processes.

  while true; do pstree -C age; sleep 2; done

Witt that you can see something like this on a 42 core system:

        ├─sshd─┬─sshd───sshd───bash───screen───screen─┬─2*[bash───vi]
        │      │                                      ├─bash───make───ansible-playboo─┬─ansible-playboo───sh───sh───python3───gmake───gmake───bash─┬─cat
        │      │                                      │                               │                                                            └─ninja─┬─sh───sphinx-build
        │      │                                      │                               │                                                                    ├─sh───c++───cc1plus
        │      │                                      │                               │                                                                    └─41*[sh───cc───cc1]
        │      │                                      │                               └─{ansible-playboo}
        │      │                                      ├─bash───pstree
        │      │                                      ├─bash───htop
        │      │                                      └─bash

Before we ere only seeing something like:

        ├─sshd─┬─sshd───sshd───bash───screen───screen─┬─bash───vi
        │      │                                      ├─bash───make───ansible-playboo─┬─ansible-playboo───sh───sh───python3───gmake───gmake───bash─┬─cat
        │      │                                      │                               │                                                            └─ninja───sh───cc───cc1
        │      │                                      │                               └─{ansible-playboo}
        │      │                                      ├─bash───pstree
        │      │                                      └─bash───htop
        │      └─sshd───sshd───bash

[0] https://docs.ansible.com/ansible/latest/collections/community/general/make_module.html
Reported-by: Adam Manzanares <a.manzanares@samsung.com>
Signed-off-by: Luis Chamberlain <mcgrof@kernel.org>